GH-381: update TC_BANK model

+++ underthesea/model_fetcher.py
@@ -19,6 +19,7 @@ class UTSModel(Enum):
     tc_general = "TC_GENERAL"
     tc_bank = "TC_BANK"
     tc_general_v131 = "TC_GENERAL_V131"
+    tc_bank_v131 = "TC_BANK_V131"
     sa_general = "SA_GENERAL"
     sa_bank = "SA_BANK"
 
@@ -61,6 +62,19 @@ class ModelFetcher:
             os.rename(
                 Path(CACHE_ROOT) / cache_dir / "tc_svm_vntc_20201228",
                 Path(CACHE_ROOT) / cache_dir / "TC_GENERAL_V131",
+            )
+            os.remove(model_path)
+
+        if model_name == "TC_BANK_V131":
+            url = "https://github.com/undertheseanlp/playground/releases/download/1.3.x/tc_svm_ubc-1_20210701.zip"
+            cached_path(url, cache_dir=cache_dir)
+            model_path = Path(CACHE_ROOT) / cache_dir / "tc_svm_ubc-1_20210701.zip"
+            cache_folder = Path(CACHE_ROOT) / cache_dir
+            zip = zipfile.ZipFile(model_path)
+            zip.extractall(cache_folder)
+            os.rename(
+                Path(CACHE_ROOT) / cache_dir / "tc_svm_ubc-1_20210701",
+                Path(CACHE_ROOT) / cache_dir / "TC_BANK_V131",
             )
             os.remove(model_path)
